починил тесты
