+++ kotouslugi-api/src/main/java/ru/practice/kotouslugi/controller/PartyController.java
@@ -7,7 +7,7 @@ import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.practice.kotouslugi.model.PoliticalParty;
-import ru.practice.kotouslugi.model.PoliticalPartyDTO;
+import ru.practice.kotouslugi.model.CreatePoliticalPartyDTO;
 import ru.practice.kotouslugi.service.PoliticalPartyService;
 
 import java.util.List;
@@ -21,8 +21,9 @@ public class PartyController {
   @PostMapping(value = "/add", produces = "application/json")
   @Operation(summary = "Добавить партию", tags = {"Кошачье АПИ"}, responses = {
     @ApiResponse(responseCode = "201", description = "Created"),
-    @ApiResponse(responseCode = "403", description = "Forbidden")})
-  public ResponseEntity<PoliticalParty> addParty(@RequestBody PoliticalPartyDTO dto){
+    @ApiResponse(responseCode = "403", description = "Forbidden"),
+    @ApiResponse(responseCode = "409", description = "Conflict")})
+  public ResponseEntity<PoliticalParty> addParty(@RequestBody CreatePoliticalPartyDTO dto){
     PoliticalParty politicalPartySaved = partyService.addPoliticalParty(dto);
     return ResponseEntity.status(201).body(politicalPartySaved);
   }
