+++ kotouslugi-core/src/main/java/ru/practice/kotouslugi/model/CreatePoliticalPartyDTO.java
@@ -1,11 +1,9 @@
 package ru.practice.kotouslugi.model;
 
-import jakarta.persistence.Column;
 import lombok.Data;
 
-import java.util.Date;
 @Data
-public class PoliticalPartyDTO {
+public class CreatePoliticalPartyDTO {
   private String name;
   private String description;
   private String logoUrl;
